Constrain background cgroup to <= 20% of CPU
This change adds a cgroup constraint on background renders to prevent
them monopolizing the CPU to improve the responsiveness of foreground
tasks.

20% was picked as an arbitrary value.  This value may be adjusted by
Chrome, since the 'chronos' user has write access to the limit value.

The 20% value was chosen as a good starting point for future work, and
to get a feeling for any regessions, if any, which may result.  Ad Hoc
testing with media players and WebGL indicate that this should not be a
problem for background media, such as playing music via Pandora or
last.fm, or music videos on youtube.

BUG=chromium-os:29195
TEST=Test using autotest which artificially loads a process in the
background cgroup (to be committed).  Manual testing verified correct
operation.

Change-Id: Ib73058801c73a47ceaddbe2fb3035499dde59815
Signed-off-by: Terry Lambert <tlambert@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/30636
Reviewed-by: Dave Moore <davemoore@chromium.org>
Reviewed-by: Olof Johansson <olofj@chromium.org>