[login_manager] Add parameterizable Chrome kill timeout
session_manager now takes a command-line flag to parameterize the number of
seconds it will wait for Chrome to shut down gracefully before killing it.

Currently, this still defaults to 3s, and is overridden to 12s in response
to the 'has_hdd' USE flag that we'll be adding to appropriate boards.

BUG=chrome-os-partner:13533
TEST=unit
TEST=boot in VM with USE flag set, ensure new timeout is in place.

Change-Id: I836f073ba3a091e429dba1258d8141704d297ca0
Reviewed-on: https://gerrit.chromium.org/gerrit/32313
Tested-by: Chris Masone <cmasone@chromium.org>
Reviewed-by: Zelidrag Hornung <zelidrag@chromium.org>
Commit-Ready: Chris Masone <cmasone@chromium.org>