[login] call RequestRestart as a method
RequestRestart used to be a signal aeons ago but is now a method, so call it as
a method. Furthermore, log if it fails so we can diagnose failures in the
future.

We also should report errors to the caller of StartDeviceWipe(), so support that
as well. I'm actually not sure how the old code _compiled_, since the generated
dbus glue would have been calling us with the wrong number of arguments.

WARNING: Running this on a test device will break anything that relies on
/usr/local, most notably python!

BUG=chromium-os:34361
TEST=adhoc
After you have logged in once and logged out again,
$ dbus-send --system --print-reply --dest=org.chromium.SessionManager \
  /org/chromium/SessionManager \
  org.chromium.SessionManagerInterface.StartDeviceWipe
should print an error message telling you a user has already logged in this
boot, and there should not be a file in your stateful partition called
factory_install_reset. Reboot manually, then run _before logging in_:
$ dbus-send --system --print-reply --dest=org.chromium.SessionManager \
  /org/chromium/SessionManager \
  org.chromium.SessionManagerInterface.StartDeviceWipe
This should print 'true' right before your system reboots, and then your
stateful partition should be gone.

Change-Id: Icaca40dd6a06e1df489a6a3b7c6fa48f6abbea4b
Reviewed-on: https://gerrit.chromium.org/gerrit/33115
Commit-Ready: Elly Jones <ellyjones@chromium.org>
Reviewed-by: Elly Jones <ellyjones@chromium.org>
Tested-by: Elly Jones <ellyjones@chromium.org>