[login_manager] Generate filename for Chrome testing FIFO safely.
We were using tempnam to do this previously, which is frowned upon.
It was probably fine, since we were only using the file name, and
ensuring the file did not exist -- and we only use this during
testing.  Now, we're using safe functions.

BUG=chromium-os:35231
TEST=unit
TEST=suite:smoke
STATUS=Fixed

Change-Id: I7d7e710ecd4f1cb0595b6dc767dff318d6417290
Reviewed-on: https://gerrit.chromium.org/gerrit/36253
Tested-by: Chris Masone <cmasone@chromium.org>
Reviewed-by: Julian Pastarmov <pastarmovj@chromium.org>
Reviewed-by: Mattias Nissler <mnissler@chromium.org>
Commit-Ready: Chris Masone <cmasone@chromium.org>