Allow Xorg to startup as nonroot user
Allow the Xorg server to be run as sudo to some other user.  Presently
still using "root" user, since the other pieces of the nonroot-X story
aren't in yet, but when that happens we should just be able to change
the "XUSER" var in session_manager_setup.sh and we'll be all good to go.

Towards the nonroot-X effort.

TEST=local build, run on x86, ARM
BUG=chromium-os:388

Change-Id: Ic46af8163b724ef18e341659512af16786ffd130
Reviewed-on: https://gerrit.chromium.org/gerrit/39301
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Chris Masone <cmasone@chromium.org>
Commit-Queue: John Sheu <sheu@chromium.org>
Tested-by: John Sheu <sheu@chromium.org>