[login_manager] Add better logging around opening NSSDB
Apparently, we are still seeing keygen fail to open the user's NSSDB.
As the db is supposed to be pre-installed in the user's cryptohome as
it's made, this is surprising.  This change adds more checks for
various kinds of problems that might exist and aborts the process
before we attempt to open the nssdb at all, so we can diagnose the
issue from crash reports.

BUG=chrome-os-partner:16365
TEST=unit
TEST=http://chromegw/i/chromiumos.tryserver/builders/alex-paladin/builds/560

Change-Id: Ibe835c226bdc19dc35956a1f94b0d34524fd1776
Reviewed-on: https://gerrit.chromium.org/gerrit/41149
Tested-by: Chris Masone <cmasone@chromium.org>
Reviewed-by: Darren Krahn <dkrahn@chromium.org>
Commit-Queue: Chris Masone <cmasone@chromium.org>