Remove Sign() from PolicyKey class
PolicyKey really represents the public half of a key pair.  In retrospect,
it's weird to have it provide a Sign() method that silently goes and
finds its private half in some implicit NSSDB in order to generate
a signature.  This becomes even funkier when we move to a world where
the session_manager is tracking multiple user sessions, each with its
own NSSDB.  So, get rid of this abstraction and make code that signs
things call NssUtil methods directly.  This will ease the transition
to multiprofile.

BUG=chromium:221770
TEST=unit tests
TEST=login_*Ownership* autotests

Change-Id: I78a32cb288b9712519f0fa81d4d8760a4c576fc6
Reviewed-on: https://gerrit.chromium.org/gerrit/48394
Commit-Queue: Chris Masone <cmasone@chromium.org>
Tested-by: Chris Masone <cmasone@chromium.org>
Reviewed-by: Mattias Nissler <mnissler@chromium.org>