The screen should be 'locked' when the lock process begins
The session_manager handles browser crashes while the screen is locked
specially, in that it ends the user session instead of restarting the
browser.  This is to keep user data safe; the user has indicated that
she wants someone to have to reauthenticate to see her stuff, and if a
crash happens while the screen is locked and we just restart things,
we'd be violating that.

The old behavior was to trigger that behavior only if we'd succeeded
in putting the lock screen up.  Instead, we should trigger the behavior
as soon as we know the user asked for screen lock.

BUG=chromium:233827
TEST=unit

Change-Id: Icfadce9c51f83c29963cab93aa66781ba1aa57a5
Reviewed-on: https://gerrit.chromium.org/gerrit/48964
Commit-Queue: Chris Masone <cmasone@chromium.org>
Tested-by: Chris Masone <cmasone@chromium.org>
Reviewed-by: Daniel Erat <derat@chromium.org>