Use a single KeyGenerator, owned by SessionManagerService
Code in SessionManagerService is written assuming that there's only
one KeyGenerator object in use by the process, while the reality is
that we have one used when mitigating the loss of an owner keypair and
another used when generating an owner keypair for the first time.

Fix this up so that there's only one.

BUG=chromium:235185
TEST=unit
TEST=login_OwnershipRetaken, login_OwnershipTaken
STATUS=Fixed

Change-Id: Ida69cc03ca2cf08856c69491dc9f527f69f5e4dd
Reviewed-on: https://gerrit.chromium.org/gerrit/49097
Commit-Queue: Chris Masone <cmasone@chromium.org>
Reviewed-by: Chris Masone <cmasone@chromium.org>
Tested-by: Chris Masone <cmasone@chromium.org>