Manage ui restarting with a separate upstart job.
Sometimes a GPU hang will cause Chrome to be unable to start. In order to
mitigate these issues, we'd like to reboot the device when this happens.
A good indicator that we might be in this situation is the ui job exiting
with a certain status code. The difficulty is that, in order to take special
action on certain exit status codes, we need to take normal respawn handling
out of the hands of upstart as well.

The new per-exit-code respawning policy implemented is:
 0: Normal exit; respawn forever.
 2: (Defined in session_manager_service.h) The browser is exiting too much too
    quickly. In some cases, such as a hung GPU, rebooting the device will allow
    the system to recover. If this happens 2 times in 180 seconds, try
    a reboot. If the device has already rebooted to try to mitigate this issue
    in the last 540 seconds, stop the UI and wait for an AU.
 All nonzero: All thing being equal, respawn up to 6 times in 60 seconds,
              then stop.

CQ-DEPEND=CL:63820, CL:62660, CL:63945

BUG=chromium:261784
TEST=pkill -9 -o "^chrome$" 8 times in 3 minutes; device should reboot
TEST=pkill -9 session_manager 7 times in a minute; ui should be stopped
TEST=stop ui and start ui should work as expected

Change-Id: If13a0c274886bce45833e7f6f218bcecdf575251
Reviewed-on: https://gerrit.chromium.org/gerrit/62781
Commit-Queue: Chris Masone <cmasone@chromium.org>
Reviewed-by: Chris Masone <cmasone@chromium.org>
Tested-by: Chris Masone <cmasone@chromium.org>